[BOOKINGSG-9282][GZ] move out styling from the component
